https://github.com/jackdewinter/pymarkdown/issues/603 - refactor
