+++ publish/coverage.json
@@ -6,8 +6,8 @@
         "totalCovered": 3581
     },
     "lineLevel": {
-        "totalMeasured": 14501,
-        "totalCovered": 14501
+        "totalMeasured": 14533,
+        "totalCovered": 14533
     }
 }
