+++ pymarkdown/container_blocks/container_grab_bag.py
@@ -3,14 +3,15 @@ Class to provide a "grab bag" for commonly used properties for the Container Blo
 """
 # pylint: disable=too-many-instance-attributes
 import logging
-from typing import Any, List, Optional
+from typing import Any, Callable, List, Optional, Tuple
 
 from pymarkdown.block_quote_data import BlockQuoteData
-from pymarkdown.container_indices import ContainerIndices
+from pymarkdown.container_blocks.container_indices import ContainerIndices
 from pymarkdown.markdown_token import MarkdownToken
 from pymarkdown.parse_block_pass_properties import ParseBlockPassProperties
 from pymarkdown.parser_logger import ParserLogger
 from pymarkdown.parser_state import ParserState
+from pymarkdown.position_marker import PositionMarker
 from pymarkdown.requeue_line_info import RequeueLineInfo
 from pymarkdown.stack_token import StackToken
 
@@ -35,7 +36,31 @@ class ContainerGrabBag:
         parser_properties: ParseBlockPassProperties,
         ignore_link_definition_start: bool,
         original_line: str,
+        recurse_fn: Callable[
+            [
+                ParserState,
+                PositionMarker,
+                bool,
+                ParseBlockPassProperties,
+                int,
+                int,
+                Optional[int],
+                Optional[int],
+                Optional[str],
+            ],
+            Tuple[
+                List[MarkdownToken],
+                Optional[str],
+                Optional[int],
+                Optional[RequeueLineInfo],
+                bool,
+                bool,
+            ],
+        ],
     ) -> None:
+        # Functions
+        self.__recurse_fn = recurse_fn
+
         # Booleans
         self.__did_blank: bool = False
         self.__was_paragraph_continuation: bool = False
@@ -662,6 +687,34 @@ class ContainerGrabBag:
         if value != self.__current_container_blocks:
             self.__log_read_write_value("current_container_blocks", value)
             self.__current_container_blocks = value
+
+    def get_recurse_fn(
+        self,
+    ) -> Callable[
+        [
+            ParserState,
+            PositionMarker,
+            bool,
+            ParseBlockPassProperties,
+            int,
+            int,
+            Optional[int],
+            Optional[int],
+            Optional[str],
+        ],
+        Tuple[
+            List[MarkdownToken],
+            Optional[str],
+            Optional[int],
+            Optional[RequeueLineInfo],
+            bool,
+            bool,
+        ],
+    ]:
+        """
+        xxx
+        """
+        return self.__recurse_fn
 
     def is_leaf_tokens_empty(self) -> bool:
         """
