+++ pymarkdown/leaf_blocks/fenced_leaf_block_processor.py
@@ -6,7 +6,7 @@ from typing import List, Optional, Tuple, cast
 
 from pymarkdown.block_quote_data import BlockQuoteData
 from pymarkdown.constants import Constants
-from pymarkdown.container_helper import ContainerHelper
+from pymarkdown.container_blocks.container_helper import ContainerHelper
 from pymarkdown.inline.inline_backslash_helper import InlineBackslashHelper
 from pymarkdown.inline_markdown_token import TextMarkdownToken
 from pymarkdown.leaf_blocks.leaf_block_helper import LeafBlockHelper
