+++ pymarkdown/leaf_blocks/thematic_leaf_block_processor.py
@@ -6,7 +6,7 @@ import logging
 from typing import List, Optional, Tuple
 
 from pymarkdown.block_quote_data import BlockQuoteData
-from pymarkdown.container_helper import ContainerHelper
+from pymarkdown.container_blocks.container_helper import ContainerHelper
 from pymarkdown.leaf_markdown_token import ThematicBreakMarkdownToken
 from pymarkdown.markdown_token import MarkdownToken
 from pymarkdown.parser_helper import ParserHelper
